refactor: make @ import point to root directory
That way we can use @/src, @/scripts and (soon) @/content consistently
