+++ src/components/Logo.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/src/lib/utils";
 
 interface LogoProps {
   size?: "micro" | "small" | "medium" | "large";
