+++ src/components/SidebarContainer.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from "react";
-import DocsSidebar from "@/components/docs/core/DocsSidebar";
-import { type ProductName } from "@/lib/route-types";
-import { type DocMeta } from "@/lib/content/docs";
+import DocsSidebar from "@/src/components/docs/core/DocsSidebar";
+import { type ProductName } from "@/src/lib/route-types";
+import { type DocMeta } from "@/src/lib/content/docs";
 
 interface SidebarContainerProps {
   product: ProductName;
