+++ src/components/ui/pagination.tsx
@@ -1,9 +1,9 @@
 import * as React from "react";
 import { ChevronLeft, ChevronRight, MoreHorizontal } from "lucide-react";
 
-import { cn } from "@/lib/utils";
-import { buttonVariants } from "@/components/ui/button";
-import type { ButtonProps } from "@/components/ui/button";
+import { cn } from "@/src/lib/utils";
+import { buttonVariants } from "@/src/components/ui/button";
+import type { ButtonProps } from "@/src/components/ui/button";
 
 const Pagination = ({ className, ...props }: React.ComponentProps<"nav">) => (
   <nav
