+++ src/routes/dev/index.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from "@tanstack/react-router";
-import DevLayout from "@/components/dev/DevLayout";
-import { environment } from "@/lib/content/environment";
+import DevLayout from "@/src/components/dev/DevLayout";
+import { environment } from "@/src/lib/content/environment";
 
 export const Route = createFileRoute("/dev/")({
   component: DevIndexPage,
