+++ src/routes/dev/social-card.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute } from "@tanstack/react-router";
 import { useState, useEffect, useRef, useCallback } from "react";
-import DevLayout from "@/components/dev/DevLayout";
-import { environment } from "@/lib/content/environment";
+import DevLayout from "@/src/components/dev/DevLayout";
+import { environment } from "@/src/lib/content/environment";
 
 export const Route = createFileRoute("/dev/social-card")({
   component: SocialCardPreview,
